fix file descriptor wrangling
It turns out janet sets FD_CLOEXEC on any file it creates, which means you
can't share it with a subprocess unless you explicitly dup it. But what fd
do you dup it to? If you pick a number that has CLOEXEC on it already, that
won't work. The safest thing to do is just not dup it, and share it as-is.
Which means no CLOEXEC. Which means bypassing janet_makefile.